requirements.txt
streamlit==1.18.0
Pillow==9.5.0
langchain==0.0.297
langchain-ollama==0.0.4